Added a sample data file, and new version.
Added a sample JSON file for the program to display. This can be used however you want, you can delete it or make your own list within the file itself. The script has also been updated to ask if the list wants to be displayed at the beginning of the program.